server/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
httpx>=0.24.0
python-dotenv>=1.0.0
jieba>=0.42.1
rank-bm25>=0.2.2
